New clientlib ver

+++ library/Kaltura/Client/Type/FlavorParams.php
@@ -67,7 +67,7 @@ class FlavorParams extends \Kaltura\Client\Type\AssetParams
 		if(count($xml->height))
 			$this->height = (int)$xml->height;
 		if(count($xml->frameRate))
-			$this->frameRate = (int)$xml->frameRate;
+			$this->frameRate = (float)$xml->frameRate;
 		if(count($xml->gopSize))
 			$this->gopSize = (int)$xml->gopSize;
 		if(count($xml->conversionEngines))
@@ -185,7 +185,7 @@ class FlavorParams extends \Kaltura\Client\Type\AssetParams
 	/**
 	 * The frame rate of the Flavor Params
 	 * 	 
-	 * @var int
+	 * @var float
 	 */
 	public $frameRate = null;
